+++ rust/src/encodings/plain.rs
@@ -109,7 +109,7 @@ impl<'a, R: Read + Seek, T: NativeType> Decoder for PlainDecoder<'a, R, T> {
         }
     }
 
-    fn value(&mut self, i: usize) -> Result<Box<dyn Scalar>> {
+    fn value(&mut self, i: i32) -> Result<Box<dyn Scalar>> {
         match self._decode(i as i32, &Some(1 as i32)) {
             Ok(arr) => Ok(Box::new(PrimitiveScalar::from(Some(arr.value(0))))),
             Err(e) => Err(e),
